legacy/evar_optimization 2/requirements.txt
absl-py==2.1.0
anyio==4.3.0
arch @ file:///home/conda/feedstock_root/build_artifacts/arch-py_1704447485783/work
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astropy @ file:///croot/astropy_1697468907928/work
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
certifi @ file:///croot/certifi_1707229174982/work/certifi
cffi==1.16.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
chess==1.10.0
clarabel @ file:///home/conda/feedstock_root/build_artifacts/clarabel_1709397141376/work
cloudpickle @ file:///home/conda/feedstock_root/build_artifacts/cloudpickle_1697464713350/work
comm==0.2.2
contourpy @ file:///croot/contourpy_1700583582875/work
cvxpy @ file:///home/conda/feedstock_root/build_artifacts/cvxpy-split_1708890478100/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
ecos @ file:///croot/ecos_1677702164573/work
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.3
fonttools==4.25.0
fqdn==1.5.1
fsspec @ file:///croot/fsspec_1701286474621/work
glfw==2.7.0
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
gym @ file:///home/conda/feedstock_root/build_artifacts/gym_1673317493781/work
gym-notices @ file:///home/conda/feedstock_root/build_artifacts/gym-notices_1671541680412/work
h11==0.14.0
httpcore==1.0.5
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-json-logger==2.0.7
pytz @ file:///croot/pytz_1695131579487/work
PyYAML @ file:///croot/pyyaml_1698096049011/work
pyzmq==25.1.2
qdldl @ file:///croot/qdldl-python_1687874402357/work
referencing==0.34.0
requests @ file:///croot/requests_1707355572290/work
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
Riskfolio-Lib @ file:///home/conda/feedstock_root/build_artifacts/riskfolio-lib_1660582854670/work
rpds-py==0.18.0
scikit-learn @ file:///croot/scikit-learn_1694788527225/work
scipy @ file:///home/conda/feedstock_root/build_artifacts/scipy-split_1706041478791/work/dist/scipy-1.12.0-cp310-cp310-linux_x86_64.whl#sha256=2f498a77f15907a2dfec49ca55a78d9463>
scs @ file:///home/conda/feedstock_root/build_artifacts/scs_1707387085142/work
Send2Trash==1.8.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
statsmodels @ file:///croot/statsmodels_1689937266057/work
sympy @ file:///croot/sympy_1701397643339/work
terminado==0.18.1
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
tinycss2==1.2.1
tomli==2.0.1
torch @ file:///home/conda/feedstock_root/build_artifacts/libtorch_1706712244942/work
torchvision @ file:///home/conda/feedstock_root/build_artifacts/torchvision_1708903236844/work
tornado==6.4
traitlets==5.14.2
types-python-dateutil==2.9.0.20240316
typing_extensions @ file:///croot/typing_extensions_1705599297034/work
tzdata @ file:///croot/python-tzdata_1690578112552/work
uri-template==1.3.0
urllib3 @ file:///croot/urllib3_1707770551213/work
virtualenv==20.25.1
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
XlsxWriter @ file:///croot/xlsxw